chore: split reducer/selectos in different files to avoid cyclic imports
