+++ webapp/src/components/GoogleAnalytics/GoogleAnalytics.container.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { getAddress } from 'modules/wallet/reducer'
+import { getAddress } from 'modules/wallet/selectors'
 import GoogleAnalytics from './GoogleAnalytics'
 
 const mapState = state => {
