+++ webapp/src/components/HomePage/Map/Map.container.js
@@ -1,8 +1,8 @@
 import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
-import { getWallet } from 'modules/wallet/reducer'
-import { getParcels } from 'modules/parcels/reducer'
-import { getDistricts } from 'modules/districts/reducer'
+import { getWallet } from 'modules/wallet/selectors'
+import { getParcels } from 'modules/parcels/selectors'
+import { getDistricts } from 'modules/districts/selectors'
 import { changeRange, setLoading, selectParcel } from 'modules/ui/actions'
 import { navigateTo } from 'modules/location/actions'
 import MapComponent from './Map'
