+++ webapp/src/components/HomePage/Sidebar/Sidebar.container.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
-import { getWallet } from 'modules/wallet/reducer'
-import { isLoading } from 'modules/ui/loading/reducer'
-import { isLoading as isLoadingParcels } from 'modules/parcels/reducer'
-import { isOpen } from 'modules/ui/sidebar/reducer'
+import { getWallet } from 'modules/wallet/selectors'
+import { isLoading } from 'modules/ui/loading/selectors'
+import { isLoading as isLoadingParcels } from 'modules/parcels/selectors'
+import { isOpen } from 'modules/ui/sidebar/selectors'
 import { openSidebar, closeSidebar } from 'modules/ui/sidebar/actions'
 import Sidebar from './Sidebar'
