+++ webapp/src/modules/districts/reducer.js
@@ -27,8 +27,3 @@ export function districtsReducer(state = INITIAL_STATE, action) {
       return state
   }
 }
-
-export const getState = state => state.districts
-export const getDistricts = state => getState(state).data
-export const isLoading = state => getState(state).loading
-export const getError = state => getState(state).error
