+++ webapp/src/modules/transfer/reducer.js
@@ -41,8 +41,3 @@ export function transferReducer(state = INITIAL_STATE, action) {
       return state
   }
 }
-
-export const getState = state => state.transfer
-export const getTransfer = state => getState(state).data
-export const isLoading = state => getState(state).loading
-export const getError = state => getState(state).error
