+++ webapp/src/modules/ui/loading/selectors.js
@@ -1,0 +1,1 @@
+export const isLoading = state => state.ui.loading
