+++ webapp/src/modules/ui/map/reducer.js
@@ -37,7 +37,3 @@ export function mapReducer(state = INITIAL_STATE, action) {
       return state
   }
 }
-
-export const getState = state => state.ui.map
-export const getRange = state => getState(state).range
-export const getSelected = state => getState(state).selected
