+++ webapp/src/modules/ui/map/selectors.js
@@ -1,0 +1,3 @@
+export const getState = state => state.ui.map
+export const getRange = state => getState(state).range
+export const getSelected = state => getState(state).selected
