+++ webapp/src/modules/ui/modal/selectors.js
@@ -1,0 +1,1 @@
+export const getModal = state => state.ui.modal
