+++ webapp/src/modules/ui/reducer.js
@@ -6,12 +6,6 @@ import { modalReducer as modal } from './modal/reducer'
 import { sidebarReducer as sidebar } from './sidebar/reducer'
 import { toastReducer as toast } from './toast/reducer'
 
-export * from './loading/reducer'
-export * from './map/reducer'
-export * from './modal/reducer'
-export * from './toast/reducer'
-export * from './sidebar/reducer'
-
 export const uiReducer = combineReducers({
   loading,
   map,
