+++ webapp/src/modules/ui/sidebar/reducer.js
@@ -1,8 +1,4 @@
-import { createSelector } from 'reselect'
-import { getWallet } from 'modules/wallet/reducer'
-import { isLoading } from 'modules/address/reducer'
 import { OPEN_SIDEBAR, CLOSE_SIDEBAR } from './actions'
-import { land } from 'lib/land'
 
 const INITIAL_STATE = {
   open: false
@@ -20,29 +16,3 @@ export function sidebarReducer(state = INITIAL_STATE, action) {
       return state
   }
 }
-
-export const getState = state => state.ui.sidebar
-export const isOpen = state => getState(state).open
-export const getStats = createSelector(
-  getWallet,
-  isLoading,
-  (wallet, loading) => {
-    const balance = wallet.balance
-    const loaded = balance !== null && !loading
-    const parcels = loaded ? wallet.parcels.length : null
-    const contribDistricts = loaded ? wallet.contributions.length : null
-    const contribMana = loaded
-      ? wallet.contributions.reduce(
-          (total, contribution) =>
-            total + land.convert(contribution.land_count, 'mana'),
-          0
-        )
-      : null
-    return {
-      balance,
-      parcels,
-      contribDistricts,
-      contribMana
-    }
-  }
-)
