+++ webapp/src/modules/ui/toast/selectors.js
@@ -1,0 +1,1 @@
+export const getToasts = state => state.ui.toast
