+++ webapp/src/modules/wallet/reducer.js
@@ -1,5 +1,3 @@
-import { createSelector } from 'reselect'
-import { getAddresses } from 'modules/address/reducer'
 import {
   FETCH_WALLET_REQUEST,
   FETCH_WALLET_SUCCESS,
@@ -46,29 +44,3 @@ export function walletReducer(state = INITIAL_STATE, action) {
       return state
   }
 }
-
-export const getState = state => state.wallet
-export const getData = state => getState(state).data
-export const isLoading = state => getState(state).loading
-export const getError = state => getState(state).error
-export const getAddress = state => getData(state).address
-export const isConnected = state => !!getData(state).address
-export const getWallet = createSelector(
-  getData,
-  getAddresses,
-  (wallet, addresses) => {
-    const address = addresses[wallet.address]
-    const parcels = address ? address.parcels : []
-    const parcelsById = address ? address.parcelsById : {}
-    const contributions = address ? address.contributions : []
-    const contributionsById = address ? address.contributionsById : {}
-
-    return {
-      ...wallet,
-      parcels,
-      parcelsById,
-      contributions,
-      contributionsById
-    }
-  }
-)
